requirements.txt
# Court Data Fetcher - Python Dependencies
# Install with: pip install -r requirements.txt

# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Browser automation
selenium==4.15.2
webdriver-manager==4.0.1

# OCR and Image processing
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3

# Web scraping and HTTP requests
beautifulsoup4==4.12.2
requests==2.31.0

# Note: The following are built-in Python modules (no installation needed):
# - os, sys, time, platform, base64, io
# - pathlib, zipfile, shutil, subprocess
# - urllib.parse, winreg (Windows only)